Create MatchCardMiddle component
